Handle images that aren't 160x160 better, make build process saner
If either dimension exceeds 160 we scale it so the larger one is 160,
preserving aspect ratio. Otherwise we use the image's dimensions as is.
Black bars are avoided in the output.

Build process: the idea of injecting everything into a single file was
never particularly sane though it made deployment trivial. I abandoned
it because I think it could limit caching of the glint frames.